feat: extract feed service into dedicated module
