+++ api/src/main/java/org/open4goods/api/config/ApiConfig.java
@@ -13,11 +13,11 @@ import org.open4goods.api.services.ScrapperOrchestrationService;
 import org.open4goods.api.services.VerticalsGenerationService;
 import org.open4goods.api.services.completion.IcecatCompletionService;
 import org.open4goods.api.services.completion.ResourceCompletionService;
-import org.open4goods.api.services.feed.AbstractFeedService;
-import org.open4goods.api.services.feed.AwinFeedService;
-import org.open4goods.api.services.feed.EffiliationFeedService;
-import org.open4goods.api.services.feed.FeedConfiguration;
-import org.open4goods.api.services.feed.FeedService;
+import org.open4goods.services.feedservice.service.AbstractFeedService;
+import org.open4goods.services.feedservice.service.AwinFeedService;
+import org.open4goods.services.feedservice.service.EffiliationFeedService;
+import org.open4goods.services.feedservice.config.FeedConfiguration;
+import org.open4goods.services.feedservice.service.FeedService;
 import org.open4goods.api.services.store.DataFragmentStoreService;
 import org.open4goods.commons.helper.DevModeService;
 import org.open4goods.commons.model.constants.TimeConstants;
