+++ services/feedservice/README.md
@@ -1,0 +1,16 @@
+# Feedservice
+
+Provides utilities to load product feed definitions from affiliation providers such as Awin and Effiliation. It relies on remote file caching and serialisation services.
+
+## Configuration
+
+Feed parameters are supplied via `FeedConfiguration` instances in your application properties.
+
+## Build & Test
+
+```bash
+mvn clean install
+mvn test
+```
+
+See the [main project](../../README.md) for details.
