+++ services/feedservice/pom.xml
@@ -2,37 +2,36 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
+
   <parent>
     <groupId>org.open4goods</groupId>
     <artifactId>org.open4goods</artifactId>
     <version>0.0.1-SNAPSHOT</version>
-    <relativePath>..</relativePath>
+    <relativePath>../..</relativePath>
   </parent>
-  <artifactId>static</artifactId>
-  <name>static</name>
+
+  <artifactId>feedservice</artifactId>
 
   <dependencies>
     <dependency>
-      <groupId>org.springframework.boot</groupId>
-      <artifactId>spring-boot-starter-web</artifactId>
-      <exclusions>
-        <exclusion>
-          <groupId>org.springframework.boot</groupId>
-          <artifactId>spring-boot-starter-tomcat</artifactId>
-        </exclusion>
-      </exclusions>
+      <groupId>org.open4goods</groupId>
+      <artifactId>commons</artifactId>
+      <version>${global.version}</version>
     </dependency>
     <dependency>
-      <groupId>org.springframework.boot</groupId>
-      <artifactId>spring-boot-starter-undertow</artifactId>
+      <groupId>org.open4goods</groupId>
+      <artifactId>remotefilecaching</artifactId>
+      <version>${global.version}</version>
     </dependency>
     <dependency>
-      <groupId>org.springframework.boot</groupId>
-      <artifactId>spring-boot-starter-cache</artifactId>
+      <groupId>org.open4goods</groupId>
+      <artifactId>serialisation</artifactId>
+      <version>${global.version}</version>
     </dependency>
     <dependency>
-      <groupId>org.springframework.boot</groupId>
-      <artifactId>spring-boot-starter-security</artifactId>
+      <groupId>org.open4goods</groupId>
+      <artifactId>model</artifactId>
+      <version>${global.version}</version>
     </dependency>
     <dependency>
       <groupId>org.springframework.boot</groupId>
@@ -43,43 +42,6 @@
       <artifactId>spring-boot-configuration-processor</artifactId>
       <optional>true</optional>
     </dependency>
-
-    <dependency>
-      <groupId>org.open4goods</groupId>
-      <artifactId>commons</artifactId>
-      <version>${global.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>org.open4goods</groupId>
-      <artifactId>verticals</artifactId>
-      <version>${global.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>org.open4goods</groupId>
-      <artifactId>product-repository</artifactId>
-      <version>${global.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>org.open4goods</groupId>
-      <artifactId>remotefilecaching</artifactId>
-      <version>${global.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>org.open4goods</groupId>
-      <artifactId>favicon</artifactId>
-      <version>${global.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>org.open4goods</groupId>
-      <artifactId>serialisation</artifactId>
-      <version>${global.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>org.open4goods</groupId>
-      <artifactId>feedservice</artifactId>
-      <version>${global.version}</version>
-    </dependency>
-
     <dependency>
       <groupId>org.springframework.boot</groupId>
       <artifactId>spring-boot-starter-test</artifactId>
@@ -113,5 +75,4 @@
       </plugin>
     </plugins>
   </build>
-
 </project>
