+++ ui/pom.xml
@@ -146,6 +146,11 @@
             <artifactId>github-feedback</artifactId>
             <version>${global.version}</version>
         </dependency>
+        <dependency>
+            <groupId>org.open4goods</groupId>
+            <artifactId>feedservice</artifactId>
+            <version>${global.version}</version>
+        </dependency>
